perf: improve detection of types
